+++ src/main/java/com/biddingmate/biddinggo/payment/controller/PaymentController.java
@@ -1,14 +1,13 @@
-package com.biddingmate.biddinggo.virtualaccount.controller;
+package com.biddingmate.biddinggo.payment.controller;
 
 import com.biddingmate.biddinggo.common.response.ApiResponse;
-import com.biddingmate.biddinggo.virtualaccount.dto.CreateVirtualAccountRequest;
-import com.biddingmate.biddinggo.virtualaccount.dto.CreateVirtualAccountResponse;
-import com.biddingmate.biddinggo.virtualaccount.dto.GetVirtualAccountResponse;
-import com.biddingmate.biddinggo.virtualaccount.service.VirtualAccountService;
+import com.biddingmate.biddinggo.payment.dto.CreateVirtualAccountRequest;
+import com.biddingmate.biddinggo.payment.dto.CreateVirtualAccountResponse;
+import com.biddingmate.biddinggo.payment.dto.GetVirtualAccountResponse;
+import com.biddingmate.biddinggo.payment.service.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,27 +18,27 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/v1/virtual-accounts")
+@RequestMapping("/api/v1/payments")
 @RequiredArgsConstructor
-public class VirtualAccountController {
-    private final VirtualAccountService virtualAccountService;
+public class PaymentController {
+    private final PaymentService paymentService;
 
-    @PostMapping("")
+    @PostMapping("/virtual-accounts")
     public ResponseEntity<ApiResponse<CreateVirtualAccountResponse>> createVirtualAccount(
             // @AuthenticationPrincipal PrincipalDetails principal,
             @RequestBody CreateVirtualAccountRequest request) {
 
-        CreateVirtualAccountResponse result = virtualAccountService.createVirtualAccount(request);
+        CreateVirtualAccountResponse result = paymentService.createVirtualAccount(request);
 
         return ApiResponse.of(HttpStatus.OK, null, "가상계좌 발급 성공", result);
     }
 
-    @GetMapping("")
+    @GetMapping("virtual-accounts")
 //    public ResponseEntity<ApiResponse<List<GetVirtualAccountResponse>>> getVirtualAccount(@AuthenticationPrincipal PrincipalDetails principal) {
 //        Member member = principal.getMember();
 //        Long memberId = member.getId();
     public ResponseEntity<ApiResponse<List<GetVirtualAccountResponse>>> getVirtualAccount(@RequestParam Long memberId) {
-        List<GetVirtualAccountResponse> result = virtualAccountService.getVirtualAccount(memberId);
+        List<GetVirtualAccountResponse> result = paymentService.getVirtualAccount(memberId);
