+++ src/main/java/com/biddingmate/biddinggo/payment/model/VirtualAccount.java
@@ -1,4 +1,4 @@
-package com.biddingmate.biddinggo.virtualaccount.model;
+package com.biddingmate.biddinggo.payment.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
