+++ src/main/resources/mapper/payment/virtual-account-mapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.biddingmate.biddinggo.virtualaccount.mapper.VirtualAccountMapper">
+<mapper namespace="com.biddingmate.biddinggo.payment.mapper.VirtualAccountMapper">
 
     <insert id="insert" parameterType="VirtualAccount" useGeneratedKeys="true" keyProperty="id">
         INSERT INTO virtual_account(
